+++ ContextDesktop/src/main/java/com/github/migi_1/Context/model/entity/IMovable.java
@@ -1,5 +1,6 @@
 package com.github.migi_1.Context.model.entity;
 
+import com.github.migi_1.Context.model.entity.behaviour.MoveBehaviour;
 
 /**
  * Interface for every entity that moves, has a method getMovableBehaviour.
