+++ ContextDesktop/src/main/java/com/github/migi_1/Context/model/entity/Platform.java
@@ -2,6 +2,7 @@ package com.github.migi_1.Context.model.entity;
 
 import java.util.HashMap;
 
+import com.github.migi_1.Context.model.entity.behaviour.ConstantSpeedMoveBehaviour;
 import com.github.migi_1.Context.utility.ProjectAssetManager;
 import com.github.migi_1.ContextMessages.PlatformPosition;
 import com.jme3.math.Vector3f;
