+++ ContextDesktop/src/main/java/com/github/migi_1/Context/model/entity/behaviour/MoveBehaviour.java
@@ -1,4 +1,4 @@
-package com.github.migi_1.Context.model.entity;
+package com.github.migi_1.Context.model.entity.behaviour;
 
 import com.jme3.math.Vector3f;
