+++ ContextDesktop/src/test/java/com/git/migi_1/Context/entity/behaviour/TestStaticMoveBehaviour.java
@@ -1,11 +1,11 @@
-package com.git.migi_1.Context.entity;
+package com.git.migi_1.Context.entity.behaviour;
 
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Before;
 import org.junit.Test;
 
-import com.github.migi_1.Context.model.entity.StaticMoveBehaviour;
+import com.github.migi_1.Context.model.entity.behaviour.StaticMoveBehaviour;
 import com.jme3.math.Vector3f;
 
 /**
